base: Confirm device credentials with the proper user
Launch the confirm device credential activity in the proper user, i.e.
the current user or the parent of the specified profile. In doing so,
make sure to query for activity handlers that apply to the current user
or parent user, in the unlikely event that its handler is not the
Settings package or is otherwise different from that of the system user.

This change primarily affects the unlocking of work profiles in
secondary users.

Co-authored-by: Tommy Webb <tommy@calyxinstitute.org>
Change-Id: I7706a2ce02a71bc10c3272cdd3d5cf84e84f0a51
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Aston-Martinn <advaithbhat9@gmail.com>