base: Fix cancelDraw issue of blast-sync.
If the first relayoutWindow of VRI during blastSync does not
create a surface, then this invalid drawing updates mLastSyncId.
This will cause the second relayoutWindow to be cancelDraw, causing blastsync timeout.

So when we update mLastSyncId, make sure the rootview is visible,
which can fix the blastsync timeout issue caused by invalid relayoutWindow.

Bug: 357648052
Test: Manual test
Change-Id: Ibc7e88cd4b7f6a061913e26298335834185c4817
Signed-off-by: yangxiao <yx894373599@gmail.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Aston-Martinn <advaithbhat9@gmail.com>